Associate URNs with profiles.
Content type profiles should be specified using a URI. This makes it
possible to associate a URN (which is a URI) with profiles. Now HAL
serializers that conform to a profile with an associated URN will get
that URN as the profile parameter in the content type. If the profile
does not specify a URN then the URL will be used instead.